requirements.txt
flask
prometheus_client
requests
kubernetes
PyYAML
